Individual Subscription: add buttons
